made login and changed index
